moved num_p to library
