+++ subtract.dart
@@ -1,5 +1,5 @@
 import 'dart:math';
-import 'num_p.dart';
+import 'package:num_p/src/num_p.dart';
 import 'format.dart';
 
 /*
